Made thread quasi ID truly unique on 64-bit machines.
We used to make them fit on 32-bit quantities, but that cannot make them
totally unique among threads in a portable way.  Having them be 64-bit
quantities on 64-bit machines ensures this uniqueness and allows "same thread"
detection in thread_element_matches() when we're just one page before or
after the previously used one for the thread.

+++ src/lib/thread.h
@@ -40,7 +40,7 @@
 typedef void (*thread_pvalue_free_t)(void *value, void *arg);
 
 typedef unsigned long thread_t;
-typedef unsigned int thread_qid_t;		/* Quasi Thread ID */
+typedef size_t thread_qid_t;		/* Quasi Thread ID */
 
 #ifdef I_PTHREAD
 #include <pthread.h>
